pyproject.toml
[tool.poetry]
name = "pystq"
version = "0.0.1"
description = ""
authors = ["tosh223 <rev.to12@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
google-cloud-pubsub = "^2.4.0"
google-cloud-monitoring = "^2.0.1"
boto3 = "^1.17.18"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
